--- app/src/main/java/ideafactory/bit/agh/ddapp/sgsg.java
@@ -1,4 +1,0 @@
-package ideafactory.bit.agh.ddapp;
-
-public class sgsg {
-}
